+++ custom_components/solarcharger/config/config_options_flow.py
@@ -25,7 +25,6 @@ from ..const import (
     ENTITY_CHARGER_ON_OFF_SWITCH,
     ENTITY_CHARGER_PLUGGED_IN_SENSOR,
     ENTITY_CHARGER_SET_CHARGE_CURRENT,
-    ENTITY_CHARGER_STEP_CURRENT_LIST,
     ENTITY_DEVICE_GET_CHARGE_LIMIT,
     ENTITY_DEVICE_LOCATION_SENSOR,
     ENTITY_DEVICE_SET_CHARGE_LIMIT,
@@ -84,6 +83,7 @@ from ..const import (
     SENSOR_DELTA_ALLOCATED_POWER,
     SUBENTRY_CHARGER_TYPES,
     SWITCH_REDUCE_CHARGE_LIMIT_DIFFERENCE,
+    TEXT_CHARGER_STEP_CURRENT_LIST,
     TIME_CHARGE_ENDTIME_FRIDAY,
     TIME_CHARGE_ENDTIME_MONDAY,
     TIME_CHARGE_ENDTIME_SATURDAY,
@@ -362,7 +362,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_CHARGER_MAX_SPEED,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry, NUMBER_CHARGER_POWER_FACTOR, use_default
@@ -371,7 +371,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_CHARGER_POWER_FACTOR,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry, NUMBER_CHARGER_MIN_CURRENT, use_default
@@ -380,7 +380,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_CHARGER_MIN_CURRENT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry, NUMBER_CHARGER_MIN_WORKABLE_CURRENT, use_default
@@ -389,7 +389,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_CHARGER_MIN_WORKABLE_CURRENT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry,
@@ -400,7 +400,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_CHARGER_MIN_WORKABLE_POWER_PAUSE_THRESHOLD,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry,
@@ -411,7 +411,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_CHARGER_MIN_WORKABLE_POWER_RESUME_THRESHOLD,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry, NUMBER_CHARGER_PRIORITY, use_default
@@ -420,7 +420,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_CHARGER_PRIORITY,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry, NUMBER_CHARGER_POWER_ALLOCATION_WEIGHT, use_default
@@ -429,7 +429,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_CHARGER_POWER_ALLOCATION_WEIGHT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry, SENSOR_DELTA_ALLOCATED_POWER, use_default
@@ -438,7 +438,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 SENSOR_DELTA_ALLOCATED_POWER,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             #####################################
             # Local device entities
@@ -480,6 +480,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 ENTITY_CHARGER_ON_OFF_SWITCH,
                 SWITCH_ENTITY_SELECTOR_READ_ONLY,
                 SWITCH_ENTITY_SELECTOR,
+                overridable=True,
             ),
             self._optional(
                 subentry, ENTITY_CHARGER_CHARGING_SENSOR, use_default
@@ -504,16 +505,16 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 NUMBER_CHARGER_MAX_CURRENT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
-                subentry, ENTITY_CHARGER_STEP_CURRENT_LIST, use_default
+                subentry, TEXT_CHARGER_STEP_CURRENT_LIST, use_default
             ): choose_selector(
                 api_entities,
-                ENTITY_CHARGER_STEP_CURRENT_LIST,
+                TEXT_CHARGER_STEP_CURRENT_LIST,
                 TEXT_ENTITY_SELECTOR_READ_ONLY,
                 TEXT_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry, ENTITY_CHARGER_GET_CHARGE_CURRENT, use_default
@@ -522,6 +523,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 ENTITY_CHARGER_GET_CHARGE_CURRENT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
+                overridable=True,
             ),
             self._optional(
                 subentry, ENTITY_CHARGER_SET_CHARGE_CURRENT, use_default
@@ -530,6 +532,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 ENTITY_CHARGER_SET_CHARGE_CURRENT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
+                overridable=True,
             ),
             self._optional(
                 subentry, ENTITY_DEVICE_SOC_SENSOR, use_default
@@ -546,7 +549,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 ENTITY_DEVICE_GET_CHARGE_LIMIT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry, ENTITY_DEVICE_SET_CHARGE_LIMIT, use_default
@@ -555,7 +558,7 @@ class ConfigOptionsFlowHandler(OptionsFlow):
                 ENTITY_DEVICE_SET_CHARGE_LIMIT,
                 NUMBER_ENTITY_SELECTOR_READ_ONLY,
                 NUMBER_ENTITY_SELECTOR,
-                modifiable_if_solarcharger_entity=True,
+                overridable=True,
             ),
             self._optional(
                 subentry, ENTITY_DEVICE_LOCATION_SENSOR, use_default